refactor: move code around so the UI doesn't load huge modules as a side effect
