+++ src/libs/broadcast/broadcast.ts
@@ -8,7 +8,7 @@ import { Hex } from '../../interfaces/hex'
 import { TxnRequest } from '../../interfaces/keystore'
 import { Network } from '../../interfaces/network'
 import { RPCProvider } from '../../interfaces/provider'
-import { getSafeBroadcastTxn } from '../../libs/safe/safe'
+import { getSafeBroadcastTxn } from '../../libs/safe/helpers'
 import wait from '../../utils/wait'
 import { AccountOp, GasFeePayment, getSignableCalls } from '../accountOp/accountOp'
 import { Call } from '../accountOp/types'
