+++ src/libs/humanizer/erc7730/registry.ts
@@ -22,7 +22,7 @@ import { Message } from '../../../interfaces/userRequest'
 import { withTimeout } from '../../../utils/with-timeout'
 import { AccountOp } from '../../accountOp/accountOp'
 import { Call } from '../../accountOp/types'
-import { decodeMultiSend } from '../../safe/safe'
+import { decodeMultiSend } from '../../safe/helpers'
 import { getAbiBytesCalldataWithPadding, multiSendInterface } from './calldata'
 import { getEip712EncodeTypeHash } from './eip712'
 import { fetchRelayerResource } from './fetch'
