+++ src/libs/swapAndBridge/constants.ts
@@ -1,0 +1,10 @@
+export enum SwapAndBridgeFormStatus {
+  Empty = 'empty',
+  Invalid = 'invalid',
+  FetchingRoutes = 'fetching-routes',
+  NoRoutesFound = 'no-routes-found',
+  InvalidRouteSelected = 'invalid-route-selected',
+  ReadyToEstimate = 'ready-to-estimate',
+  ReadyToSubmit = 'ready-to-submit',
+  Proceeded = 'proceeded'
+}
