+++ src/libs/tracer/accessListCall.ts
@@ -12,7 +12,7 @@ import { Network } from '../../interfaces/network'
 import { getRpcProvider } from '../../services/provider'
 import { BaseAccount } from '../account/BaseAccount'
 import { AccountOp } from '../accountOp/accountOp'
-import { encodeCalls } from '../safe/safe'
+import { encodeCalls } from '../safe/helpers'
 
 const safeSimulateTxAccessorAbi = [
   'function simulate(address to, uint256 value, bytes data, uint8 operation)'
